requirements.txt
pydantic>=2.12.5
google-genai>=1.0.0
pinecone>=5.0.0
httpx>=0.27.0
scikit-learn>=1.5.0
networkx>=3.0
numpy>=2.0.0
openai>=1.0.0
python-dotenv>=1.0.0
google-search-results>=2.4.2
Pillow>=10.0.0
